187 xarray datatree (#193)
* Feature/datatree (#185)

* adding a datatree method to be able to open an entire gribfile instead of having to supply filters

* Add additional human-readable names to LEVEL_NAME_MAPPING

* Add tests for DataTree structure and filtering in GRIB2 files

* remove accidental submodule

* remove datatree interp test

* require xarray >= 2023.4.0

* make datatree optional

* Add DataTree support checks and refactor DataTree accessor registration

* increase datatree xarray requirement

* Add additional level mappings for cloud layers in LEVEL_NAME_MAPPING

* Fix syntax error in LEVEL_NAME_MAPPING by adding missing comma

* Adding a custom level naming table.

Adding "table_4_5_grib2io_level_name". This is a table that maps
code values from table 4.5 to a custom name. The name is either
mapped directly to a CF-conventions standard name or is a custom
name for grib2io.

These names are used to define the DataTree.

* Update for xarray backend

Level names now being pulled from new custom grib2io table for
GRIB2 table 4.5.

* Renaming xarray datatree test.

[skip ci]

* Update _grib2io.py

Add to_section3() method for Grib2GridDef class.

* Update tests/test_to_grib2.py

Remove dtype attr for leadTime before decode.  This is a temporary
change.

* Update Grib2GridDef and some code clean up.

Clean up of Grib2GridDef property returns and ProjParameters class.

* Update build_linux.yml to gcc-14.

* Update license in pyproject.toml

Modify setup.py with test prints to diagnose finding omp.h

* Update setup.py

* Update setup.py

Added find_openmp_include()

* Update setup.py

* Update setup.py

Clean up of setup.py

* Clean up setup.py

[skip ci]

* Update macOS testing to gcc-14.

* Update setup.py

Make sure full omp include path is returned

Minor update to GRIB2 tables.

---------

+++ src/grib2io/templates.py
@@ -651,50 +651,50 @@ class ProjParameters:
         projparams['a'] = 1.0
         projparams['b'] = 1.0
         if obj.earthRadius is not None:
-            projparams['a'] = obj.earthRadius
-            projparams['b'] = obj.earthRadius
+            projparams['a'] = float(obj.earthRadius)
+            projparams['b'] = float(obj.earthRadius)
         else:
-            if obj.earthMajorAxis is not None: projparams['a'] = obj.earthMajorAxis
-            if obj.earthMajorAxis is not None: projparams['b'] = obj.earthMinorAxis
+            if obj.earthMajorAxis is not None: projparams['a'] = float(obj.earthMajorAxis)
+            if obj.earthMajorAxis is not None: projparams['b'] = float(obj.earthMinorAxis)
         if obj.gdtn == 0:
             projparams['proj'] = 'longlat'
         elif obj.gdtn == 1:
             projparams['o_proj'] = 'longlat'
             projparams['proj'] = 'ob_tran'
-            projparams['o_lat_p'] = -1.0*obj.latitudeSouthernPole
-            projparams['o_lon_p'] = obj.anglePoleRotation
-            projparams['lon_0'] = obj.longitudeSouthernPole
+            projparams['o_lat_p'] = float(-1.0*obj.latitudeSouthernPole)
+            projparams['o_lon_p'] = float(obj.anglePoleRotation)
+            projparams['lon_0'] = float(obj.longitudeSouthernPole)
         elif obj.gdtn == 10:
             projparams['proj'] = 'merc'
-            projparams['lat_ts'] = obj.latitudeTrueScale
-            projparams['lon_0'] = 0.5*(obj.longitudeFirstGridpoint+obj.longitudeLastGridpoint)
+            projparams['lat_ts'] = float(obj.latitudeTrueScale)
+            projparams['lon_0'] = float(0.5*(obj.longitudeFirstGridpoint+obj.longitudeLastGridpoint))
         elif obj.gdtn == 20:
             if obj.projectionCenterFlag == 0:
                 lat0 = 90.0
             elif obj.projectionCenterFlag == 1:
                 lat0 = -90.0
             projparams['proj'] = 'stere'
-            projparams['lat_ts'] = obj.latitudeTrueScale
+            projparams['lat_ts'] = float(obj.latitudeTrueScale)
             projparams['lat_0'] = lat0
-            projparams['lon_0'] = obj.gridOrientation
+            projparams['lon_0'] = float(obj.gridOrientation)
         elif obj.gdtn == 30:
             projparams['proj'] = 'lcc'
-            projparams['lat_1'] = obj.standardLatitude1
-            projparams['lat_2'] = obj.standardLatitude2
-            projparams['lat_0'] = obj.latitudeTrueScale
-            projparams['lon_0'] = obj.gridOrientation
+            projparams['lat_1'] = float(obj.standardLatitude1)
+            projparams['lat_2'] = float(obj.standardLatitude2)
+            projparams['lat_0'] = float(obj.latitudeTrueScale)
+            projparams['lon_0'] = float(obj.gridOrientation)
         elif obj.gdtn == 31:
             projparams['proj'] = 'aea'
-            projparams['lat_1'] = obj.standardLatitude1
-            projparams['lat_2'] = obj.standardLatitude2
-            projparams['lat_0'] = obj.latitudeTrueScale
-            projparams['lon_0'] = obj.gridOrientation
+            projparams['lat_1'] = float(obj.standardLatitude1)
+            projparams['lat_2'] = float(obj.standardLatitude2)
+            projparams['lat_0'] = float(obj.latitudeTrueScale)
+            projparams['lon_0'] = float(obj.gridOrientation)
         elif obj.gdtn == 40:
             projparams['proj'] = 'eqc'
         elif obj.gdtn == 32769:
             projparams['proj'] = 'aeqd'
-            projparams['lon_0'] = obj.longitudeCenterGridpoint
-            projparams['lat_0'] = obj.latitudeCenterGridpoint
+            projparams['lon_0'] = float(obj.longitudeCenterGridpoint)
+            projparams['lat_0'] = float(obj.latitudeCenterGridpoint)
         return projparams
     def __set__(self, obj, value):
         raise RuntimeError
